[BEAM-5495] Use ClassGraph to detect classpath resources
ClassGraph seems to be able to detect the needed resources for all java
versions (both pre and post 8) and includes JPMS support. This is why it
is a good fit for the job. Other than that using it significantly
reduces code that otherwise would have to be maintained.

See: https://github.com/classgraph/classgraph for more info.